move saved playbooks to a separate client module so only part of the home page is rendered client-side
